refactor: move widgets to shared dir, move tests out of printer directory
